requirements.txt
flask
pandas
numpy
scikit-learn
matplotlib
psutil
fpdf
requests
pyobjc-framework-AppKit; platform_system == "Darwin"
flask-cors
